Python/requirements.txt
pyserial>=3.5
matplotlib>=3.7
numpy>=1.24
pandas>=2.0
